requirements.txt
streamlit
pydantic
requests
python-dotenv
